+++ spacecat/helpers/perms.py
@@ -1,6 +1,8 @@
 import configparser
 import os
+import re
 import sqlite3
+import timeit
 
 import discord
 from discord.ext import commands
@@ -23,81 +25,117 @@ def new(guild):
         config['PermsGroups'][str(guild.default_role.id)] = userperms
         config['PermsUsers'] = {}
 
+
 def check():
     def predicate(ctx):
-        # If user is the server administrator, always allow
-        command_parents = ctx.command.qualified_name.split(' ')
-
+        """
+        Checks if the user has permission to use the command based on
+        if they are a server admin, if they have the required
+        permission on the server, or if it's a default permission set
+        within the bot's global config.
+        """
+        # Allow if user is a server administrator
         if ctx.author.guild_permissions.administrator:
             return True
 
-        # Open server's database file
-        db = sqlite3.connect(settings.data + 'spacecat.db')
-        cursor = db.cursor()
-
-        # Check if specific user has a permission in server
-        queries = []
-        queries.append((ctx.guild.id, ctx.author.id, "*"))
-        queries.append((ctx.guild.id, ctx.author.id, f"{ctx.command.cog.qualified_name}.*"))
-
-        # Check subcommand parents for permission
-        perm = ''
-        for index, command in enumerate(command_parents):
-            if index == 0:
-                perm = command
-            else:
-                perm = f"{perm}.{command}"
-            queries.append((ctx.guild.id, ctx.author.id, f"{ctx.command.cog.qualified_name}.{perm}.*"))
-        queries.append((ctx.guild.id, ctx.author.id, f"{ctx.command.cog.qualified_name}.{perm}%"))
+        # Grab useful command variables
+        module = ctx.command.cog.qualified_name
+        command = ctx.command.qualified_name
+        command_values = command.split(' ')
         
-        # Execute user perm queries
-        for query in queries:
+        # Add queries for the global, module, and command wildcards,
+        # plus the command on its own
+        perm = command_values[0]
+        checks = {'*', f'{module}.*', f'{module}.{perm}.*'}
+        regex_query = re.compile('Preset.[a-zA-Z0-9]+')
+        for command_value in command_values[1:]:
+            perm = f"{perm}.{command_value}"
+            checks.add(f'{module}.{perm}.*')
+        checks.add(f'{module}.{perm}')
+        
+        # Query database to allow if user has the required permission
+        db = sqlite3.connect(settings.data + 'spacecat.db')
+        db.row_factory = lambda cursor, row: row[0]
+        cursor = db.cursor()
+        query = (ctx.guild.id, ctx.author.id)
+        cursor.execute(
+            'SELECT permission FROM user_permission '
+            'WHERE server_id=? AND user_id=?', query)
+        user_results = cursor.fetchall()
+        presets = list(filter(regex_query.match, user_results))
+        if set(user_results).intersection(checks):
+            return True
+
+        # Query database to allow if any group that the user is assigned
+        # to has the required permission
+        for group in ctx.author.roles:
+            query = (ctx.guild.id, group.id)
             cursor.execute(
-                'SELECT permission FROM user_permission WHERE server_id=? AND user_id=? AND permission LIKE ?', query)
-            check = cursor.fetchall()
-            if check:
+                'SELECT permission FROM group_permission '
+                'WHERE server_id=? AND group_id=?', query)
+            group_results = cursor.fetchall()
+            presets = presets + list(filter(regex_query.match, group_results))
+            if set(group_results).intersection(checks):
                 return True
 
-        # Convert query from user ID to every group ID that a user has
-        group_queries = []
-        for role in ctx.author.roles:
-            for index, query in enumerate(queries):
-                query_conversion = list(query)
-                query_conversion[1] = role.id
-                group_queries.append(query_conversion)
-
-        # Execute all group perm queries
-        for query in group_queries:
-            cursor.execute(
-                'SELECT permission FROM group_permission WHERE server_id=? AND group_id=? AND permission LIKE ?', query)
-            check = cursor.fetchall()
-            query = (ctx.guild.id, role.id)
-            parent_check = parent_perms(ctx, cursor, query)
-            if check or parent_check:
+            # Execute recurring parent check
+            parent_query = (ctx.guild.id, query[1])
+            parent_check, presets = parent_perms(
+                ctx, cursor, parent_query, checks, presets)
+            if parent_check:
                 return True
 
-    def parent_perms(ctx, cursor, query):
+        # Query permission presets from config that the user or group may have
+        config = toml.load(settings.data + 'config.toml')
+        for preset in presets:
+            comparison = list(
+                set(config['permissions'][preset[7:]]).intersection(checks))
+            if comparison:
+                return True
+
+        # Query config's default permission list
+        query = (ctx.guild.id,)
+        cursor.execute(
+            'SELECT advanced_permission FROM server_settings '
+            'WHERE server_id=?', query)
+        advanced = cursor.fetchone()
+        if advanced == 0:
+            comparison = set(
+                config['permissions']['default']).intersection(checks)
+            if comparison:
+                return True
+
+    def parent_perms(ctx, cursor, parent_query, checks, presets):
+        """
+        Recursively checks all parents until either all dead ends have
+        been reached, or the appropriate permission has been found.
+        """
         # Check if group has parents
         cursor.execute(
-                'SELECT parent_id FROM group_parent WHERE server_id=? AND child_id=?', query)
+            'SELECT parent_id FROM group_parent '
+            'WHERE server_id=? AND child_id=?', parent_query)
         parents = cursor.fetchall()
 
         # Check parent groups for permission
-        if parents:
-            for parent in parents:
-                query = (ctx.guild.id, parent[0], f"{ctx.command.cog.qualified_name}.{ctx.command.name}")
-                cursor.execute(
-                    'SELECT permission FROM group_permission WHERE server_id=? AND group_id=? AND permission LIKE ?', query)
-                check = cursor.fetchall()
-                if check:
-                    return True
+        for parent in parents:
+            perm_query = (ctx.guild.id, parent)
+            cursor.execute(
+                'SELECT permission FROM group_permission '
+                'WHERE server_id=? AND group_id=?', perm_query)
+            results = cursor.fetchall()
+            regex_query = re.compile('Preset.[a-zA-Z0-9]+')
+            presets = presets + list(filter(regex_query.match, results))
+            if set(results).intersection(checks):
+                return True, presets
 
-                query = (ctx.guild.id, parent[0])
-                parent_check = parent_perms(ctx, cursor, f"{ctx.command.cog.qualified_name}.{ctx.command.name}")
-                if parent_check:
-                    return True
+            # Check next parent level
+            new_parent_query = (ctx.guild.id, parent)
+            parent_check, presets = parent_perms(
+                ctx, cursor, new_parent_query, checks, presets)
+            if parent_check:
+                return True, presets
+        return False, presets
 
-        
     return commands.check(predicate)